isolate chakra breakpoint detection into a separate hook
